requirements.txt
captum==0.3.1
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dstain @ file:///home/groups/jamesz/bryanhe/digital-staining
Flask==1.1.2
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5
numba==0.53.0
numpy==1.20.1
opencv-python==4.5.1.48
openslide-python==1.1.2
Pillow==8.1.2
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2021.3.17
torch==1.8.0
torchvision==0.9.0
tqdm==4.59.0
typing-extensions==3.7.4.3
umap-learn==0.5.1
Werkzeug==1.0.1
